test(e2e): migrate remaining Perps Detox smokes to Appium
Port the last Detox Perps smoke specs to Appium, delete tests/smoke/perps,
remove Detox SmokePerps CI jobs, and shard Appium Perps into 2 splits.
