+++ .github/workflows/run-appium-smoke-tests-android.yml
@@ -89,15 +89,15 @@ jobs:
       }}
     strategy:
       matrix:
-        split: [1]
+        split: [1, 2]
       fail-fast: false
     uses: ./.github/workflows/run-appium-e2e-workflow.yml
     with:
-      test-suite-name: appium-perps-android-smoke
+      test-suite-name: appium-perps-android-smoke-${{ matrix.split }}
       platform: android
       test_suite_tag: SmokePerps
       split_number: ${{ matrix.split }}
-      total_splits: 1
+      total_splits: 2
       test-timeout-minutes: 25
       build_type: ${{ inputs.build_type }}
       metamask_environment: ${{ inputs.metamask_environment }}
