+++ .github/workflows/run-appium-smoke-tests-ios.yml
@@ -68,15 +68,15 @@ jobs:
       }}
     strategy:
       matrix:
-        split: [1]
+        split: [1, 2]
       fail-fast: false
     uses: ./.github/workflows/run-appium-e2e-workflow.yml
     with:
-      test-suite-name: appium-perps-ios-smoke
+      test-suite-name: appium-perps-ios-smoke-${{ matrix.split }}
       platform: ios
       test_suite_tag: SmokePerps
       split_number: ${{ matrix.split }}
-      total_splits: 1
+      total_splits: 2
       test-timeout-minutes: 25
       build_type: ${{ inputs.build_type || 'main' }}
       metamask_environment: ${{ inputs.metamask_environment || 'e2e' }}
