+++ .github/workflows/run-e2e-smoke-tests-android.yml
@@ -24,23 +24,6 @@ permissions:
   id-token: write
 
 jobs:
-  perps-android-smoke:
-    if: contains(fromJson(inputs.selected_tags), 'ALL') || contains(fromJson(inputs.selected_tags), 'SmokePerps')
-    strategy:
-      matrix:
-        split: [1]
-      fail-fast: false
-    uses: ./.github/workflows/run-e2e-workflow.yml
-    with:
-      test-suite-name: perps-android-smoke-${{ matrix.split }}
-      platform: android
-      test_suite_tag: 'SmokePerps'
-      split_number: ${{ matrix.split }}
-      total_splits: 1
-      changed_spec_files: ${{ inputs.changed_spec_files }}
-      runner_provider: ${{ inputs.runner_provider }}
-    secrets: inherit
-
   # SmokeWalletPlatform Detox job removed — tag fully migrated to Appium
   # (run-appium-smoke-tests-android.yml / appium-wallet-platform-android-smoke).
 
@@ -126,7 +109,6 @@ jobs:
         )
       }}
     needs:
-      - perps-android-smoke
       - network-abstraction-android-smoke
       - network-expansion-android-smoke
       - confirmations-android-smoke
