+++ .github/workflows/run-e2e-smoke-tests-ios.yml
@@ -42,24 +42,6 @@ jobs:
       runner_provider: ${{ inputs.runner_provider }}
     secrets: inherit
 
-  perps-ios-smoke:
-    if: contains(fromJson(inputs.selected_tags), 'ALL') || contains(fromJson(inputs.selected_tags), 'SmokePerps')
-    strategy:
-      matrix:
-        split: [1]
-      fail-fast: false
-    uses: ./.github/workflows/run-e2e-workflow.yml
-    with:
-      test-suite-name: perps-ios-smoke-${{ matrix.split }}
-      platform: ios
-      test_suite_tag: 'SmokePerps'
-      split_number: ${{ matrix.split }}
-      total_splits: 1
-      changed_spec_files: ${{ inputs.changed_spec_files }}
-      build_type: 'main'
-      runner_provider: ${{ inputs.runner_provider }}
-    secrets: inherit
-
   # SmokeWalletPlatform Detox job removed — tag fully migrated to Appium
   # (run-appium-smoke-tests-ios.yml / appium-wallet-platform-ios-smoke).
 
@@ -132,7 +114,6 @@ jobs:
       }}
     needs:
       - confirmations-ios-smoke
-      - perps-ios-smoke
       - network-abstraction-ios-smoke
       - network-expansion-ios-smoke
       - multichain-api-ios-smoke
