+++ tests/smoke-appium/perps/perps-position-liquidation.spec.ts
@@ -1,0 +1,91 @@
+import { test as appiumTest } from '../../framework/fixtures/playwright/index.js';
+import { SmokePerps } from '../../tags.js';
+import { withFixtures } from '../../framework/fixtures/FixtureHelper.js';
+import PerpsMarketDetailsView from '../../page-objects/Perps/PerpsMarketDetailsView.js';
+import PerpsE2EModifiers from '../../helpers/perps/perps-modifiers.js';
+import { TestSuiteParams, Utilities } from '../../framework/index.js';
+import CommandQueueServer from '../../framework/fixtures/CommandQueueServer.js';
+import { openPosition } from '../../flows/perps.flow.js';
+import {
+  beginPerpsSmokeTestPlaywright,
+  buildPerpsSmokeFixture,
+  PERPS_SMOKE_MARKET_SYMBOL,
+  PERPS_SMOKE_PERMISSIONS,
+  setupPerpsSmokeMocks,
+} from '../../helpers/perps/perps-smoke-helpers.js';
+
+const NON_LIQUIDATING_MARK_PRICE = '2400.00';
+const LIQUIDATING_MARK_PRICE = '1.00';
+
+const queueLiquidationCheckAtPrice = async (
+  commandQueueServer: CommandQueueServer,
+  symbol: string,
+  price: string,
+) => {
+  await PerpsE2EModifiers.updateMarketPriceServer(
+    commandQueueServer,
+    symbol,
+    price,
+  );
+  await PerpsE2EModifiers.triggerLiquidationServer(commandQueueServer, symbol);
+};
+
+const waitForCommandQueueToProcess = async (
+  commandQueueServer: CommandQueueServer,
+) => {
+  commandQueueServer.requestStateExport();
+  await commandQueueServer.getExportedState();
+};
+
+appiumTest.describe(SmokePerps('Perps Position Liquidation'), () => {
+  appiumTest(
+    'liquidates a long position when mark price moves below liquidation price',
+    async ({ driver: _driver, currentDeviceDetails }) => {
+      await withFixtures(
+        {
+          fixture: buildPerpsSmokeFixture(),
+          restartDevice: true,
+          currentDeviceDetails,
+          permissions: PERPS_SMOKE_PERMISSIONS,
+          testSpecificMock: setupPerpsSmokeMocks,
+          useCommandQueueServer: true,
+        },
+        async ({ commandQueueServer }: TestSuiteParams) => {
+          if (!commandQueueServer) {
+            throw new Error('Command queue server not found');
+          }
+
+          await beginPerpsSmokeTestPlaywright();
+
+          await openPosition(PERPS_SMOKE_MARKET_SYMBOL, 'long');
+
+          await queueLiquidationCheckAtPrice(
+            commandQueueServer,
+            PERPS_SMOKE_MARKET_SYMBOL,
+            NON_LIQUIDATING_MARK_PRICE,
+          );
+          await waitForCommandQueueToProcess(commandQueueServer);
+          await PerpsMarketDetailsView.expectClosePositionButtonVisible();
+
+          await queueLiquidationCheckAtPrice(
+            commandQueueServer,
+            PERPS_SMOKE_MARKET_SYMBOL,
+            LIQUIDATING_MARK_PRICE,
+          );
+
+          await Utilities.executeWithRetry(
+            async () => {
+              await PerpsMarketDetailsView.expectClosePositionButtonNotVisible();
+            },
+            {
+              interval: 1000,
+              timeout: 30000,
+              description:
+                'wait for Close position to disappear after liquidation',
+            },
+          );
+        },
+      );
+    },
+  );
+});
