--- tests/smoke/perps/perps-position.spec.ts
@@ -1,91 +1,0 @@
-import { loginToApp } from '../../flows/wallet.flow';
-import { withFixtures } from '../../framework/fixtures/FixtureHelper';
-import { SmokePerps } from '../../tags';
-import FixtureBuilder from '../../framework/fixtures/FixtureBuilder';
-import { navigateToPerpsOrderEntry } from '../../flows/perps.flow';
-import {
-  PERPS_ARBITRUM_MOCKS,
-  mockPerpsGeolocation,
-} from '../../api-mocking/mock-responses/perps-arbitrum-mocks';
-import { RampsRegions, RampsRegionsEnum } from '../../framework/Constants';
-import PerpsMarketDetailsView from '../../page-objects/Perps/PerpsMarketDetailsView';
-import PerpsOrderView from '../../page-objects/Perps/PerpsOrderView';
-import PerpsView from '../../page-objects/Perps/PerpsView';
-import { createLogger, LogLevel } from '../../framework';
-import { Mockttp } from 'mockttp';
-import { setupRemoteFeatureFlagsMock } from '../../api-mocking/helpers/remoteFeatureFlagsHelper';
-
-// E2E environment setup - mocks auto-configure via hasTestOverrides flag
-
-const logger = createLogger({
-  name: 'PerpsPositionSpec',
-  level: LogLevel.INFO,
-});
-
-describe(SmokePerps('Perps Position'), () => {
-  it('opens a long position with custom profit and closes it', async () => {
-    await withFixtures(
-      {
-        fixture: new FixtureBuilder()
-          .withPerpsProfile('no-positions')
-          .withPerpsFirstTimeUser(false)
-          .withAccountTreeController()
-          .withNetworkController({
-            type: 'rpc',
-            chainId: '0xa4b1',
-            rpcUrl: 'https://arb1.arbitrum.io/rpc',
-            nickname: 'Arbitrum One',
-            ticker: 'ETH',
-          })
-          .withTokensForAllPopularNetworks([
-            {
-              address: '0xff970a61a04b1ca14834a43f5de4533ebddb5cc8',
-              symbol: 'USDC',
-              decimals: 6,
-              name: 'USD Coin',
-              type: 'erc20',
-            },
-          ])
-          .build(),
-        restartDevice: true,
-        permissions: { notifications: 'YES' },
-        testSpecificMock: async (mockServer: Mockttp) => {
-          await setupRemoteFeatureFlagsMock(mockServer, {});
-          await PERPS_ARBITRUM_MOCKS(mockServer);
-          await mockPerpsGeolocation(
-            mockServer,
-            RampsRegions[RampsRegionsEnum.SPAIN],
-          );
-        },
-      },
-      async () => {
-        logger.info('💰 Using E2E mock balance - no wallet import needed');
-        logger.info('🎯 Mock account: $10,000 total, $8,000 available');
-        await loginToApp();
-        // Keep consistency with other passing Perps smoke flows where
-        // streaming/network activity can block Detox idling.
-        await device.disableSynchronization();
-
-        await navigateToPerpsOrderEntry('ETH', 'long');
-
-        // Custom TP trigger above mock ETH mark (~2500) for a long
-        await PerpsOrderView.tapTakeProfitButton();
-        await PerpsOrderView.enterCustomTakeProfitTriggerPrice('2800');
-        await PerpsOrderView.tapPlaceOrderButton();
-
-        // Wait for screen ready and assert Close Position availability
-        await PerpsMarketDetailsView.waitForScreenReady();
-        await PerpsMarketDetailsView.expectClosePositionButtonVisible();
-
-        await PerpsView.tapClosePositionButton();
-
-        logger.info('📉 E2E Mock: Preparing to close position');
-
-        await PerpsView.tapClosePositionBottomSheetButton();
-
-        logger.info('🎉 E2E Mock: Position closed successfully');
-        logger.info('💰 E2E Mock: Balance updated with P&L');
-      },
-    );
-  });
-});
